requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
Appium-Python-Client==5.2.4
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
pillow==12.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PySocks==1.7.1
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
